updated resource file for new dialog defines
git-svn-id: svn://svn.code.sf.net/p/ditto-cp/code/trunk@587 595ec19a-5cb4-439b-94a8-42fb3063c22c

+++ CopyProperties.cpp
@@ -223,19 +223,19 @@ void CCopyProperties::OnOK()
 			{
 				LoadDataIntoCClip(clip);
 
+				if(CheckGlobalHotKey(clip) == FALSE)
+				{
+					if(MessageBox(_T("Error registering global hot key\n\nContinue?"), _T(""), MB_YESNO|MB_ICONWARNING) == IDNO)
+					{
+						return;
+					}
+				}
+
 				if(clip.ModifyMainTable())
 				{
 					if(m_bDeletedData)
 					{    
 						DeleteFormats(m_lCopyID, m_DeletedData);
-					}
-
-					if(CheckGlobalHotKey(clip) == FALSE)
-					{
-						if(MessageBox(_T("Error registering global hot key\n\nContinue?"), _T(""), MB_YESNO|MB_ICONWARNING) == IDNO)
-						{
-							return;
-						}
 					}
 				}
 			}
